1、版本号升级至2.1.002 2、实现xxx材料表变更流程

+++ code/code/borui/app/企业版/MainProgram/model/初始化/业务初始化/InitMateriel.cs
@@ -288,7 +288,7 @@ namespace MainProgram.model
                         {
                             InitMaterielTable record = new InitMaterielTable();
                             record.materielID = Convert.ToInt32(materielID.ToString());
-                            record.value = Convert.ToInt32(value.ToString());
+                            record.value = Convert.ToDouble(value.ToString());
                             record.price = Convert.ToDouble(price.ToString());
 
                             insert(record, false);
